+++ src/Utils/CategoryUtils.php
@@ -6,6 +6,7 @@ declare(strict_types=1);
 
 namespace Basecom\FixturePlugin\Utils;
 
+use Shopware\Core\Content\Category\CategoryCollection;
 use Shopware\Core\Content\Category\CategoryEntity;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
@@ -14,6 +15,9 @@ use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 
 readonly class CategoryUtils
 {
+    /**
+     * @param EntityRepository<CategoryCollection> $categoryRepository
+     */
     public function __construct(
         private EntityRepository $categoryRepository,
     ) {