+++ src/Utils/CustomerUtils.php
@@ -8,10 +8,14 @@ use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\System\Salutation\SalutationCollection;
 use Shopware\Core\System\Salutation\SalutationEntity;
 
 readonly class CustomerUtils
 {
+    /**
+     * @param EntityRepository<SalutationCollection> $salutationRepository
+     */
     public function __construct(
         private EntityRepository $salutationRepository,
     ) {